backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.8",
    "django-allauth>=65.13.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-simple-history>=3.10.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "networkx>=3.6",
    "pillow>=12.0.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "svgwrite>=1.4.3",
]
